refactor: router, remove lazy loading
