+++ packages/client/src/app/index.ts
@@ -1,2 +1,3 @@
-export { router } from './router/routes';
+export { routes } from './router/routes';
+export { router } from './router/router';
 export { store } from './store/store';
