+++ packages/client/src/app/router/router.ts
@@ -1,0 +1,5 @@
+import { createBrowserRouter } from 'react-router-dom';
+
+import { routes } from './routes';
+
+export const router = createBrowserRouter(routes);
