+++ packages/client/src/pages/forum-topic-edit/index.ts
@@ -1,2 +1,3 @@
-export { ForumTopicEditPageAsync as ForumTopicEditPage } from './ui/ForumTopicEditPage.async';
+export { ForumTopicEditPageAsync } from './ui/ForumTopicEditPage.async';
+export { ForumTopicEditPage } from './ui/ForumTopicEditPage';
 export { topicLoader } from './model/loader';
