+++ packages/client/src/pages/forum-topic-view/index.ts
@@ -1,1 +1,2 @@
-export { ForumTopicViewPageAsync as ForumTopicViewPage } from './ui/ForumTopicViewPage.async';
+export { ForumTopicViewPageAsync } from './ui/ForumTopicViewPage.async';
+export { ForumTopicViewPage } from './ui/ForumTopicViewPage';
