+++ packages/client/src/pages/home-layout/index.ts
@@ -1,1 +1,2 @@
-export { HomeLayoutAsync as HomeLayout } from './ui/HomeLayout.async';
+export { HomeLayoutAsync } from './ui/HomeLayout.async';
+export { HomeLayout } from './ui/HomeLayout';
