+++ packages/client/src/pages/home/index.ts
@@ -1,1 +1,2 @@
-export { HomePageAsync as HomePage } from './ui/HomePage.async';
+export { HomePageAsync } from './ui/HomePage.async';
+export { HomePage } from './ui/HomePage';
