+++ packages/client/src/pages/login/index.ts
@@ -1,1 +1,2 @@
-export { LoginPageAsync as LoginPage } from './ui/LoginPage.async';
+export { LoginPageAsync } from './ui/LoginPage.async';
+export { LoginPage } from './ui/LoginPage';
